[api] Move QgsGeometry enums to Qgis, promote to enum classes

+++ src/core/settings/qgssettingsentry.h
@@ -895,7 +895,7 @@ class CORE_EXPORT QgsSettingsEntryEnumFlag : public QgsSettingsEntryBase
      * \note for Python bindings, a custom implementation is achieved in Python directly
      */
     QgsSettingsEntryEnumFlag( const QString &key, QgsSettings::Section section, const T &defaultValue, const QString &description = QString() )
-      : QgsSettingsEntryBase( key, section, QMetaEnum::fromType<T>().isFlag() ? QVariant( QMetaEnum::fromType<T>().valueToKeys( defaultValue ) ) : QVariant( QMetaEnum::fromType<T>().valueToKey( defaultValue ) ), description )
+      : QgsSettingsEntryBase( key, section, QMetaEnum::fromType<T>().isFlag() ? QVariant( QMetaEnum::fromType<T>().valueToKeys( static_cast <int >( defaultValue ) ) ) : QVariant( QMetaEnum::fromType<T>().valueToKey( static_cast< int >( defaultValue ) ) ), description )
     {
       mMetaEnum = QMetaEnum::fromType<T>();
       Q_ASSERT( mMetaEnum.isValid() );
@@ -990,10 +990,10 @@ class CORE_EXPORT QgsSettingsEntryEnumFlag : public QgsSettingsEntryBase
 
       if ( !mMetaEnum.isFlag() )
       {
-        const char *enumKey = mMetaEnum.valueToKey( value );
+        const char *enumKey = mMetaEnum.valueToKey( static_cast< int >( value ) );
         if ( enumKey == nullptr )
         {
-          QgsDebugMsg( QStringLiteral( "Invalid enum value '%1'." ).arg( value ) );
+          QgsDebugMsg( QStringLiteral( "Invalid enum value '%1'." ).arg( static_cast< int >( value ) ) );
           return false;
         }
 
@@ -1001,10 +1001,10 @@ class CORE_EXPORT QgsSettingsEntryEnumFlag : public QgsSettingsEntryBase
       }
       else
       {
-        QByteArray flagKeys = mMetaEnum.valueToKeys( value );
+        QByteArray flagKeys = mMetaEnum.valueToKeys( static_cast< int >( value ) );
         if ( flagKeys.isEmpty() )
         {
-          QgsDebugMsg( QStringLiteral( "Invalid flag value '%1'." ).arg( value ) );
+          QgsDebugMsg( QStringLiteral( "Invalid flag value '%1'." ).arg( static_cast< int >( value ) ) );
           return false;
         }
         return QgsSettingsEntryBase::setVariantValue( flagKeys, dynamicKeyPartList );
